requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-allauth==0.51.0
django-ckeditor==6.5.0
django-countries==7.3.2
django-debug-toolbar==3.6.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
